add some ifs

+++ install.sh
@@ -4,7 +4,7 @@ CYAN='\033[0;36m'
 NORMAL='\033[0m'
 YELLOW='\033[0;33m'
 FILESPATH=`pwd -P`/files
-BACKUPPATH="${HOME}/backup"
+BACKUPPATH="$HOME/backup"
 
 printf "${GREEN}"
 echo "______      _    __ _ _           "
@@ -22,33 +22,41 @@ fi
 printf "${GREEN}DONE!${NORMAL}\n"
 
 printf "${CYAN}Making backup folder...\n${NORMAL}"
-if [ ! -d $BACKUPPATH ]; then
+if [ ! -d "$BACKUPPATH" ]; then
   mkdir ${BACKUPPATH}
 fi
 printf "${GREEN}DONE!${NORMAL}\n"
 
 printf "${CYAN}Backing things up...\n${NORMAL}"
-if [ ! -L "${HOME}/.tmux.conf" ]; then
+if [ ! -L "$HOME/.tmux.conf" ] && [ -f "$HOME/.tmux.conf" ]; then
   mv --backup=numbered ~/.tmux.conf ${BACKUPPATH}/.tmux.conf.backup
 fi
 
-if [ ! -L "${HOME}/.config/nvim/init.vim" ]; then
+if [ ! -L "$HOME/.config/nvim/init.vim" ] && [ -f "$HOME/.config/nvim/init.vim" ]; then
   mv --backup=numbered ~/.config/nvim/init.vim ${BACKUPPATH}/init.vim.backup
 fi
 
-if [ ! -L "${HOME}/.zshrc" ]; then
+if [ ! -L "$HOME/.zshrc" ] && [ -f "$HOME/.zshrc" ]; then
   mv --backup=numbered ~/.zshrc ${BACKUPPATH}/.zshrc.backup
 fi
 printf "${GREEN}DONE!${NORMAL}\n"
 
 printf "${CYAN}Install oh-my-zsh and plugins...${NORMAL}\n"
-sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
-git clone git://github.com/zsh-users/zsh-autosuggestions $HOME/.oh-my-zsh/custom/plugins/zsh-autosuggestions
-git clone https://github.com/bhilburn/powerlevel9k.git $HOME/.oh-my-zsh/custom/themes/powerlevel9k
+if [ ! -f "$HOME/.oh-my-zsh/tools/install.sh" ]; then
+  sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
+fi
+if [ ! -d "$HOME/.oh-my-zsh/custom/plugins/zsh-autosuggestions" ]; then
+  git clone git://github.com/zsh-users/zsh-autosuggestions ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions
+fi
+if [ ! -d "$HOME/.oh-my-zsh/custom/themes/powerlevel9k/" ]; then
+  git clone https://github.com/bhilburn/powerlevel9k.git ~/.oh-my-zsh/custom/themes/powerlevel9k
+fi
 printf "${GREEN}DONE!${NORMAL}\n"
 
 printf "${CYAN}Install vim plug...${NORMAL}\n"
-curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+if [ ! -f "$HOME/.config/nvim/autoload/plug.vim" ]; then
+  curl -fLo ~/.config/nvim/autoload/plug.vim --create-dirs https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+fi
 printf "${GREEN}DONE!${NORMAL}\n"
 
 printf "${CYAN}Creating symlinks...${NORMAL}\n"
@@ -58,18 +66,16 @@ ln -fs ${FILESPATH}/.zshrc ~/.zshrc
 printf "${GREEN}DONE!${NORMAL}\n"
 
 printf "${CYAN}Install fonts...${NORMAL}\n"
-if [ ! -d "~/.local/share/fonts/" ]; then
+if [ ! -d "$HOME/.local/share/fonts/" ]; then
   mkdir -p ~/.local/share/fonts/
 fi
 
 if [ ! -f "$HOME/.local/share/fonts/DejaVu Sans Mono Nerd Font Complete.ttf" ]; then
   curl -fLo ~/.local/share/fonts/"DejaVu Sans Mono Nerd Font Complete.ttf" https://github.com/ryanoasis/nerd-fonts/raw/master/patched-fonts/DejaVuSansMono/Regular/complete/DejaVu%20Sans%20Mono%20Nerd%20Font%20Complete.ttf
 fi
-
 printf "${GREEN}DONE!${NORMAL}\n"
 
 printf "${CYAN}Changing user shell...${NORMAL}\n"
 chsh -s $(which zsh)
 printf "${GREEN}DONE!${NORMAL}\n"
-
 printf "${GREEN}COMPLETE!${NORMAL}\n"
